LayoutTests/accessibility/ios-simulator/button-with-aria-haspopup-role-expected.txt

This tests the platform attributes exposed for buttons with aria-haspopup

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS element.hasAttribute('aria-haspopup') is true
test1 has trait haspopup = true with aria-haspopup attribute = dialog
test2 has trait haspopup = true with aria-haspopup attribute = grid
test3 has trait haspopup = true with aria-haspopup attribute = listbox
test4 has trait haspopup = true with aria-haspopup attribute = menu
test5 has trait haspopup = true with aria-haspopup attribute = tree
test6 has trait haspopup = false with aria-haspopup attribute = null
test7 has trait haspopup = false with aria-haspopup attribute = false
test8 has trait haspopup = false with aria-haspopup attribute = foo
test9 has trait haspopup = false with aria-haspopup attribute = 
PASS successfullyParsed is true

TEST COMPLETE
